cmd/docker: make feature detection lazy again
Commit 20ba591b7f36c2ab24e4b3c78aaeae2836d2aea4 fixed incorrect feature
detection in the CLI, but introduced a regression; previously the "ping"
would only be executed if needed (see b39739123b845f872549e91be184cc583f5b387c),
but by not inlining the call to `ServerInfo()` would now always be called.

This patch inlines the code again to only execute the "ping" conditionally,
which allows it to be executed lazily (and omitted for commands that don't
require a daemon connection).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>