Added missing backslash to documentation sites cli snippet
I think the cli code block misses a backslash to brevent line break when copy/pasting it to a terminal.
I doubt that this is intentional, if it is, feel free to reject the pr.

Signed-off-by: Julian <gitea+julian@ic.thejulian.uk>
(cherry picked from commit 2d86b79def935af49b30ab01c7b38b0875d918ab)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>