cli/command/container: remove unused NetworkDisabled field
This comment was added in https://github.com/moby/moby/commit/7929888214741c4ab194c44e0b14ac08aca06556
when this code was still in the Moby repository. That comment doesn't appear
to apply to the CLI's usage of this struct though, as nothing in the CLI
sets this field (or uses it), so this should be safe to remove.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>